refactor: create InMemoryRoutingState instaed of RoutingState.of
